Fix(mason-null-ls): Updated configuration
Sometime since I configured my null-ls setup with mason until today
there was a breaking change in mason-null-ls concerning the way you
configure your formatters using it as the primary source of truth. This
commit addresses it.

Not that it matter anymore since null-ls has been archived and will
stop working as soon as Neovim is updated probably lol