Fix(zshrc): Fix dotfilesDiff Alias
For some reason the previous one stopped working. I wrapped the new
alias in a anonymous function because they are called automatically and
scope the local variables so they won't spill in the session.

Changed Neovim command flag syntax from -c to +"", easier to read.

Also, had to cd into the home directory instead of opening a buffer in
it in order to make it work again.